pyproject.toml
[project]
name = "fast-light-chat"
version = "0.1.0"
description = "Real-time chat support platform like Channel Talk"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # FastAPI & ASGI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Socket.IO
    "python-socketio>=5.11.0",
    # Database - PostgreSQL
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Database - MongoDB
    "motor>=3.6.0",
    # Database - Redis
    "redis>=5.2.0",
    # Authentication & Security
    "pyjwt>=2.9.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.17",
    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    # Utilities
    "python-dateutil>=2.9.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
